refactor(gui-v2): import composables from `#imports` path
