+++ packages/nc-gui-v2/components/smartsheet-toolbar/FieldListAutoCompleteDropdown.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { MetaInj } from '~/components'
+import { MetaInj } from '~/context'
 
 interface Props {
   modelValue?: string
