+++ packages/nc-gui-v2/components/smartsheet/Sidebar.vue
@@ -3,7 +3,7 @@ import { ViewTypes } from 'nocodb-sdk'
 import type { TableType } from 'nocodb-sdk'
 import type { Ref } from 'vue'
 import { inject, ref } from '#imports'
-import { ActiveViewInj, MetaInj, ViewListInj } from '~/components'
+import { ActiveViewInj, MetaInj, ViewListInj } from '~/context'
 import useViews from '~/composables/useViews'
 import { viewIcons } from '~/utils/viewUtils'
 import MdiPlusIcon from '~icons/mdi/plus'
