+++ packages/nc-gui-v2/composables/useColumn.ts
@@ -1,6 +1,6 @@
 import type { ColumnType } from 'nocodb-sdk'
 import { SqlUiFactory, UITypes, isVirtualCol } from 'nocodb-sdk'
-import useProject from '~/composables/useProject'
+import { useProject } from '#imports'
 
 export default (column: ColumnType) => {
   const { project } = useProject()
